instrument + accuracy checks ag-grid and events working

+++ app/core/projects/components/project-detail/project-detail-tab-instruments.js
@@ -14,6 +14,14 @@ var tab_instruments = ['$scope', '$timeout','$routeParams', 'SubprojectService',
         scope.selectedInstrument = null; //what they've selected in the dropdown to add to the project
 
 
+        var CheckMethodRenderer = function (param) {
+            return DataGradeMethods[param.data.CheckMethod]; //DataGradeMethods defined in config.js
+        };
+
+        var CheckDateRenderer = function (param) {
+            return moment(param.data.CheckDate).format('L');
+        };
+
         var EditLinksTemplate = function (param) {
 
             var div = document.createElement('div');
@@ -21,7 +29,7 @@ var tab_instruments = ['$scope', '$timeout','$routeParams', 'SubprojectService',
             var editBtn = document.createElement('a'); editBtn.href = '#'; editBtn.innerHTML = 'Edit';
             editBtn.addEventListener('click', function (event) {
                 event.preventDefault();
-                scope.editHabitatSubproject(param.data);
+                scope.editInstrument(param.data);
             });
             div.appendChild(editBtn);
             div.appendChild(document.createTextNode("|"));
@@ -29,7 +37,7 @@ var tab_instruments = ['$scope', '$timeout','$routeParams', 'SubprojectService',
             var delBtn = document.createElement('a'); delBtn.href = '#'; delBtn.innerHTML = 'Delete';
             delBtn.addEventListener('click', function (event) {
                 event.preventDefault();
-                scope.removeHabitatSubproject(param.data);
+                scope.removeInstrument(param.data);
             });
             div.appendChild(delBtn);
             div.appendChild(document.createTextNode("|"));
@@ -37,15 +45,69 @@ var tab_instruments = ['$scope', '$timeout','$routeParams', 'SubprojectService',
             var addBtn = document.createElement('a'); addBtn.href = '#'; addBtn.innerHTML = 'Add AC';
             addBtn.addEventListener('click', function (event) {
                 event.preventDefault();
-                scope.openHabitatItemForm(param.data, {});
+                scope.openAccuracyCheckForm(param.data, {});
             });
             div.appendChild(addBtn);
 
             return div;
         };
 
+        var EditDetailLinksTemplate = function (param) {
+
+            var div = document.createElement('div');
+
+            var editBtn = document.createElement('a'); editBtn.href = '#'; editBtn.innerHTML = 'Edit';
+            editBtn.addEventListener('click', function (event) {
+                event.preventDefault();
+                scope.openAccuracyCheckForm(getById(scope.project.Instruments, param.data.InstrumentId), param.data);
+            });
+            div.appendChild(editBtn);
+            div.appendChild(document.createTextNode("|"));
+
+            var delBtn = document.createElement('a'); delBtn.href = '#'; delBtn.innerHTML = 'Delete';
+            delBtn.addEventListener('click', function (event) {
+                event.preventDefault();
+                scope.removeAccuracyCheck(param.data);
+            });
+            div.appendChild(delBtn);
+            div.appendChild(document.createTextNode("|"));
+
+            var addBtn = document.createElement('a'); addBtn.href = '#'; addBtn.innerHTML = 'Add';
+            addBtn.addEventListener('click', function (event) {
+                event.preventDefault();
+                scope.openAccuracyCheckForm(getById(scope.project.Instruments, param.data.InstrumentId), {});
+            });
+            div.appendChild(addBtn);
+
+            return div;
+        };
+
+        //Instrument Accuracy Checks
+        var instrDetailColDefs = [
+            { cellRenderer: EditDetailLinksTemplate, width: 100, menuTabs: [] },
+            { field: 'CheckDate', headerName: 'Check Date', width: 100, sort: 'desc', cellRenderer: CheckDateRenderer, menuTabs: []},
+            { field: 'CheckMethod', headerName: 'Check Method', cellRenderer: CheckMethodRenderer, width: 230, menuTabs: ['filterMenuTab'] },
+            { field: 'Bath1Grade', headerName: 'Bath 1 Grade', width: 100, menuTabs: ['filterMenuTab'] },
+            { field: 'Bath2Grade', headerName: 'Bath 2 Grade', width: 100, menuTabs: ['filterMenuTab'] },
+            { field: 'Comments', headerName: 'Comments', width: 275, menuTabs: [] },
+        ];
+
+        scope.instrDetailGridOptions = {
+                enableSorting: true,
+                enableFilter: true,
+                enableColResize: true,
+                columnDefs: instrDetailColDefs,
+        };
+
         ///////////////instruments grid
         scope.instrGridOptions = {
+            masterDetail: true,
+            detailCellRendererParams: {
+                detailGridOptions: scope.instrDetailGridOptions,
+                getDetailRowData: function (params) {
+                    params.successCallback(params.data.AccuracyChecks);
+                },
+            },
             enableSorting: true,
             enableFilter: true,
             enableColResize: true,
@@ -61,15 +123,14 @@ var tab_instruments = ['$scope', '$timeout','$routeParams', 'SubprojectService',
             selectedItems: [],
             columnDefs:
             [
-                { cellRenderer: EditLinksTemplate, width: 80 },
-                { field: 'Name', headerName: 'Name', width: 250, sort: 'asc' },
-                { field: 'SerialNumber', headerName: 'SerialNumber', width: 120 },
-                //{ field: 'Description', headerName: 'Description', width: 250 },
-                { field: 'Manufacturer', headerName: 'Manufacturer', width: 150 },
-                { field: 'Model', headerName: 'Model', width: 150 },
+                { cellRenderer: EditLinksTemplate, width: 120, menuTabs: [], },
+                { field: 'Name', headerName: 'Name', width: 250, sort: 'asc', menuTabs: ['filterMenuTab'], filter: 'text', cellRenderer: 'group' },
+                { field: 'SerialNumber', headerName: 'SerialNumber', width: 120, menuTabs: ['filterMenuTab'], filter: 'text'},
+                { field: 'Manufacturer', headerName: 'Manufacturer', width: 150, menuTabs: ['filterMenuTab'], },
+                { field: 'Model', headerName: 'Model', width: 150, menuTabs: ['filterMenuTab'], },
+                { field: 'InstrumentType.Name', headerName: 'Type', menuTabs: ['filterMenuTab'], },
+                //{ field: 'OwningDepartment.Name', headerName: 'Owner', width: 250, menuTabs: ['filterMenuTab'], },
                 
-                { field: 'OwningDepartment.Name', headerName: 'Owner', width: 250 },
-                { field: 'InstrumentType.Name', headerName: 'Type' },
                 
             ]
         };
@@ -173,24 +234,61 @@ var tab_instruments = ['$scope', '$timeout','$routeParams', 'SubprojectService',
             var promise = ProjectService.saveProjectInstrument(scope.project.Id, Instruments[0]);
 
             promise.$promise.then(function () {
-                scope.reloadProject();
+                //scope.reloadProject();
             });
         };
 
 
-        scope.removeViewInstrument = function () {
+        scope.removeInstrument = function (instrument) {
+            scope.viewInstrument = instrument;
             if (!scope.viewInstrument)
                 return;
 
-            var promise = ProjectService.removeProjectInstrument(scope.project.Id, scope.viewInstrument.Id);
+            if (confirm("Are you sure you want to remove this Instrument from this Project?")) {
+                var promise = ProjectService.removeProjectInstrument(scope.project.Id, scope.viewInstrument.Id);
 
-            promise.$promise.then(function () {
-                scope.reloadProject();
-            });
+                promise.$promise.then(function () {
+                    scope.project.Instruments.forEach(function (item, index) {
+                        if (item.Id === scope.viewInstrument.Id) {
+                            scope.project.Instruments.splice(index, 1);
+                            scope.instrGridOptions.api.setRowData(scope.project.Instruments);
+                        }
+                    });
+                });
+            }
         };
 
 
-        scope.editViewInstrument = function () {
+        scope.removeAccuracyCheck = function (ac) {
+
+            var instrument = getById(scope.project.Instruments, ac.InstrumentId);
+            if (typeof instrument === 'undefined') {
+                alert("Cannot remove that Accuracy Check. Can't find related instrument. Please share this error with your administrator.");
+                return;
+            }
+
+            if (confirm("Are you sure you want to remove this Accuracy Check from this Instrument?")) {
+                
+                var promise = ProjectService.removeInstrumentAccuracyCheck(ac.InstrumentId, ac);
+
+                promise.$promise.then(function () {
+                    instrument.AccuracyChecks.forEach(function (item, index) {
+                        if (item.Id === ac.Id) {
+                            instrument.AccuracyChecks.splice(index, 1);
+                            scope.instrGridOptions.api.setRowData(scope.project.Instruments);
+                            var the_node = scope.expandById(ac.InstrumentId);
+                            if (the_node != null)
+                                scope.instrGridOptions.api.ensureNodeVisible(the_node);
+                        }
+                    });
+                });
+            }
+        };
+
+        scope.editInstrument = function (instrument) {
+
+            scope.viewInstrument = instrument;
+
             var modalInstance = $modal.open({
                 templateUrl: 'app/core/common/components/modals/templates/modal-create-instrument.html',
                 controller: 'ModalCreateInstrumentCtrl',
@@ -198,9 +296,100 @@ var tab_instruments = ['$scope', '$timeout','$routeParams', 'SubprojectService',
             });
         };
 
+        //fired after a user saves a new or edited instrument
+        // we update the item in project's instruments then refresh the grid.
+        scope.postSaveInstrumentUpdateGrid = function (the_promise) {
+            //console.log("ok - we saved so update the grid...");
+            var total = scope.project.Instruments.length;
+            var count = 0;
+            var updated = false;
+            scope.project.Instruments.forEach(function (item, index) {
+                if (item.Id === the_promise.Id) {
+                    updated = true;
+
+                    //console.log("ok we found a match! -- updating! before:");
+                    //console.dir(scope.subprojectList[index]);
+
+                    if (the_promise.AccuracyChecks !== undefined)
+                        delete the_promise.AccuracyChecks; //remove this before the copy.
+
+                    angular.extend(scope.project.Instruments[index], the_promise); //replace the data for that item
+                    //console.log("ok we found a match! -- updating! after:");
+                    scope.instrGridOptions.api.redrawRows();
+                    console.log("done reloading grid after editing an instrument.");
+                }
+                count++;
+                if (count == total && updated == false) //if we get all done and we never found it, lets add it to the end.
+                {
+                    //console.log("ok we found never a match! -- adding!");
+                    the_promise.AccuracyChecks = [];
+                    scope.project.Instruments.push(the_promise); //add that item
+                    scope.instrGridOptions.api.setRowData([]);
+                    scope.instrGridOptions.api.setRowData(scope.project.Instruments);
+
+                    console.log("done reloading grid after adding an instrument.");
+                }
+            });
+
+            console.log("updated the list and the grid... now refreshing the instrument lists");
+            //scope.refreshSubprojectLists(); //funders, collaborators, etc.
+
+        };
 
 
-        scope.openAccuracyCheckForm = function (ac_row) {
+        //returns the (last) node or null if none found.
+        scope.expandById = function (id_in) {
+            var the_node = null;
+            scope.instrGridOptions.api.forEachNode(function (node) {
+                if (node.data.Id === id_in) {
+                    //console.log("Expanding! " + id_in);
+                    node.setExpanded(true);
+                    the_node = node;
+                }
+            });
+            return the_node;
+        };
+
+
+
+
+        //called by the modal once the instrument accuracy check is successfully saved.
+        scope.postInstrumentAccuracyCheckUpdateGrid = function (edited_item) {
+            //edit our instrument's accuracy check and then reload the grid.
+            var edited = false;
+            scope.project.Instruments.forEach(function (item, index) {
+                if (item.Id === edited_item.InstrumentId) {
+                    item.AccuracyChecks.forEach(function (instr_item, instr_item_index) {
+                        if (instr_item.Id === edited_item.Id) {
+                            angular.extend(instr_item, edited_item); //replace the data for that item
+                            console.log("OK!! we edited that accuracy check item");
+                            edited = true;
+                        }
+                    });
+                    if (!edited) {
+                        item.AccuracyChecks.push(edited_item);
+                        console.log("OK we added that accuracy check item!");
+                    }
+                }
+            });
+            
+
+            scope.instrGridOptions.api.setRowData(scope.project.Instruments);
+
+            //after we setRowData, the grid collapses our expanded item. we want it to re-expand that item and make sure it is visible.
+            var the_node = scope.expandById(edited_item.InstrumentId);
+            if (the_node != null)
+                scope.instrGridOptions.api.ensureNodeVisible(the_node);
+
+            console.log("done reloading grid after editing accuracy check item.");
+
+        };
+
+        scope.openAccuracyCheckForm = function (a_instrument, ac_row) {
+
+            scope.viewInstrument = a_instrument;
+            console.dir(a_instrument);
+
             if (ac_row)
                 scope.ac_row = ac_row;
             else
